deleted the pagination.js file and renamed data.js
